requirements.txt
cohere
faiss-cpu
langchain
langchain-community
langchain-openai
langchain_postgres
numpy
openai
pandas
pgvector
psycopg
psycopg2
psycopg2-binary
pymupdf
python-dotenv
python-Levenshtein
PyPDF2
scikit-learn
slack-sdk
slack_bolt
slackeventsapi
SQLAlchemy
tiktoken